+++ src/constants/routes.jsx
@@ -1,8 +1,4 @@
 export default {
-  //stacks
-  SplashStack: 'SplashStack',
-
-  //screens
-  Splash: 'Splash',
-  Home: 'Home',
+  Splash: "Splash",
+  Home: "Home",
 };
