+++ src/navigations/HomeStack.jsx
@@ -1,7 +1,10 @@
+import { memo } from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { Home } from "../screens";
 import { ROUTES } from "../constants";
 const Stack = createNativeStackNavigator();
+
+export default memo(HomeStack);
 
 export const HomeStack = () => {
   return (
@@ -14,4 +17,4 @@ export const HomeStack = () => {
       <Stack.Screen name={ROUTES.Home} component={Home} />
     </Stack.Navigator>
   );
-};
+}
