+++ src/navigations/SplashStack.jsx
@@ -1,9 +1,12 @@
+import { memo } from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { Splash } from "../screens";
 import { ROUTES } from "../constants";
 const Stack = createNativeStackNavigator();
 
-export const SplashStack = () => {
+export default memo(SplashStack);
+
+const SplashStack = () => {
   return (
     <Stack.Navigator
       screenOptions={{
